Added Daily Notifications
UnlockListenerStarter (BroadcastReceiver for on boot) starts up the ScreenUnlockListener service (also started by SleepTrackerGUI if it hasn't been started).  Both check to see if the service is started by using a static instance that ScreenUnlockListener has.  The ScreenUnlockListener service starts a BroadcastReceiver to check for screen unlocks called UnlockScreenReceiver.  A screen unlock calls a createNotification method in ScreenUnlockListener which checks if it is after 10 a.m and before 10 p.m, and if the notification has not yet been sent that day, and if both are true, uses SleepTracker to fetch the sleep info, and then create a notification informing the user of their sleep deprivation.  Clicking on that notification will bring a user into the app.

In the future there is probably a better way to do this using alarms to start up and close down the receiver.  Other things to be added is actually changing how sleep deprivation is calculated based on the mode, and adding a settings area where users can change their currently selected mode.